az204/01_storage/cosmos-db/requirements.txt
azure-cosmos
python-dotenv
aiohttp